Project Set Up Complete with Data Science Section
The backend folder has been reformatted to use a more traditional flask file structure. The frontend folder contains the code from the frontend branch merged into this branch, which contains a basic menu format. The added data_analysis folder contains a data folder that is to be explored from the Poppycox Clover system, a folder to contain notebooks for EDA and model training/evaluation, and a folder to contain exported ML models.